added infinite support and disk full support

+++ disk_full_test.sh
@@ -11,22 +11,27 @@ mkdir $WORKSPACE
 echo "Which directory do you want to check for integrity? (Directory must exist)"
 read CHECK_DIR
 
+echo "What block size (bytes) do you want to test with?"
+read BLOCK_SIZE
+
 #CHECK_DIR=/mnt/smallusb
 
-#echo "Inifinite (1) or until disk full(2) ?"
-#read INFINITE
+echo "Inifinite (1) or until disk full(2) ?"
+read INFINITE
 
+echo "removing old data from checked directory"
 rm -rf $CHECK_DIR/sample_data/
 mkdir $CHECK_DIR/sample_data/
 
 #create sample file of 40mb
-dd if=/dev/urandom of=$WORKSPACE/sample bs=40000000 count=1
+echo "generating sample file"
+dd if=/dev/urandom of=$WORKSPACE/sample bs=BLOCK_SIZE count=1
 
 #start infinite loop that copies sample files to directed disk
  SUCCESS=0
  FAILS=0
- COUNTER=1
-         while [  $COUNTER > 0 ]; do
+ COUNTER=0
+         while [  $COUNTER -ge 0 ]; do
          	cp $WORKSPACE/sample $CHECK_DIR/sample_data/$COUNTER.sample 
          	if [ $? -eq 0 ]; then
          		cmp --silent $WORKSPACE/sample $CHECK_DIR/sample_data/$COUNTER.sample && let SUCCESS=SUCCESS+1  || let FAILS=FAILS+1 
@@ -37,8 +42,13 @@ dd if=/dev/urandom of=$WORKSPACE/sample bs=40000000 count=1
 			 	if [ $usep -ge 99 ]; then
 			 		#enable to do disk ful write
 			 		cmp --silent $WORKSPACE/sample $CHECK_DIR/sample_data/0.sample && echo "TEST SUCCESS $COUNTER files, Index 0 still in tact and disk full \n" >> diskfull.log  || echo "TEST FAIL $COUNTER files, Index 0 corrupted when disk full \n" >> diskfull.log
-			 		rm -rf $CHECK_DIR/sample_data/
-					mkdir $CHECK_DIR/sample_data/
+			 		echo "disk full, starting over"
+			 		if [ $INFINITE -eq 1 ]; do 
+			 			rm -rf $CHECK_DIR/sample_data/
+						mkdir $CHECK_DIR/sample_data/
+					else
+						exit
+					fi
 				fi
 			fi
 			done
